fix(ui): fix infinite API calls in Jellyfin grid
additional refactoring to help other needless re-renders
