+++ web/src/components/channel_config/ProgrammingSelector.tsx
@@ -1,6 +1,3 @@
-import { useJellyfinGenres, useJellyfinUserLibraries } from '@/hooks/jellyfin/useJellyfinApi.ts';
-import type { JellyfinGenresResponse, JellyfinItem } from '@tunarr/types/jellyfin';
-import { useKnownMedia } from '@/store/programmingSelector/selectors.ts';
 import {
   Alert,
   Box,
@@ -11,45 +8,26 @@ import {
   Stack,
   Typography,
 } from '@mui/material';
-import { tag } from '@tunarr/types';
-import { isPlexDirectory } from '@tunarr/types/plex';
-import {
-  capitalize,
-  find,
-  first,
-  isEmpty,
-  isNil,
-  isUndefined,
-  map,
-  sortBy,
-} from 'lodash-es';
+import { capitalize, find, isEmpty, isUndefined, map } from 'lodash-es';
 import { useCallback, useEffect, useState } from 'react';
 import { Emby, Jellyfin, Plex } from '../../helpers/constants.ts';
-import { sortEmbyLibraries } from '../../helpers/embyUtil.ts';
-import { sortJellyfinLibraries } from '../../helpers/jellyfinUtil.ts';
-import { useEmbyUserLibraries } from '../../hooks/emby/useEmbyApi.ts';
-import {
-  usePlexLibraries,
-  usePlexPlaylists,
-} from '../../hooks/plex/usePlex.ts';
 import { useMediaSources } from '../../hooks/settingsHooks.ts';
 import { useCustomShows } from '../../hooks/useCustomShows.ts';
 import { useProgrammingSelectionContext } from '../../hooks/useProgrammingSelectionContext.ts';
 import { Route } from '../../routes/channels_/$channelId/programming/add.tsx';
 import useStore from '../../store/index.ts';
 import {
-  addKnownMediaForEmbyServer,
-  addKnownMediaForJellyfinServer,
-  addKnownMediaForPlexServer,
   setProgrammingListLibrary,
   setProgrammingListingServer,
-  setProgrammingGenre,
 } from '../../store/programmingSelector/actions.ts';
 import { ProgramViewToggleButton } from '../base/ProgramViewToggleButton.tsx';
 import { AddMediaSourceButton } from '../settings/media_source/AddMediaSourceButton.tsx';
 import { CustomShowProgrammingSelector } from './CustomShowProgrammingSelector.tsx';
+import { EmbyLibrarySelector } from './emby/EmbyLibrarySelector.tsx';
 import { EmbyProgrammingSelector } from './emby/EmbyProgrammingSelector.tsx';
+import { JellyfinLibrarySelector } from './jellyfin/JellyfinLibrarySelector.tsx';
 import { JellyfinProgrammingSelector } from './jellyfin/JellyfinProgrammingSelector.tsx';
+import { PlexLibrarySelector } from './plex/PlexLibrarySelector.tsx';
 import PlexProgrammingSelector from './plex/PlexProgrammingSelector.tsx';
 
 type Props = {
@@ -68,49 +46,12 @@ export const ProgrammingSelector = ({
     useMediaSources();
   const selectedServer = useStore((s) => s.currentMediaSource);
   const selectedLibrary = useStore((s) => s.currentMediaSourceView);
-  const selectedGenre = useStore((s) => s.currentMediaGenre);
-  const knownMedia = useKnownMedia();
   const [mediaSource, setMediaSource] = useState(selectedServer?.name);
   const navigate = Route.useNavigate();
-
-  // Convenience sub-selectors for specific library types
-  const selectedPlexLibrary =
-    selectedLibrary?.type === Plex ? selectedLibrary.view : undefined;
-  const selectedJellyfinLibrary =
-    selectedLibrary?.type === Jellyfin ? selectedLibrary.view : undefined;
-  const selectedEmbyLibrary =
-    selectedLibrary?.type === Emby ? selectedLibrary.view : undefined;
-
   const viewingCustomShows = mediaSource === 'custom-shows';
 
-  const { data: plexLibraryChildren } = usePlexLibraries(
-    selectedServer?.id ?? tag(''),
-    selectedServer?.type === Plex,
-  );
-
-  const { data: plexPlaylists, isLoading: plexPlaylistsLoading } =
-    usePlexPlaylists(
-      selectedServer?.id ?? tag(''),
-      selectedServer?.type === Plex,
-    );
-
-  const { data: jellyfinLibraries } = useJellyfinUserLibraries(
-    selectedServer?.id ?? '',
-    selectedServer?.type === Jellyfin,
-  );
-
-  const { data: embyLibraries } = useEmbyUserLibraries(
-    selectedServer?.id ?? '',
-    selectedServer?.type === Emby,
-  );
-
-  const { data: jellyfinGenres } = useJellyfinGenres(
-    selectedServer?.id ?? tag(''),
-    selectedJellyfinLibrary?.Id ?? '',
-    selectedServer?.type === Jellyfin && !!selectedJellyfinLibrary,
-  ) as { data?: JellyfinGenresResponse };
-
   useEffect(() => {
+    console.log('105');
     const server =
       !isUndefined(mediaSources) && !isEmpty(mediaSources)
         ? (find(mediaSources, ({ id }) => id === initialMediaSourceId) ??
@@ -119,79 +60,6 @@ export const ProgrammingSelector = ({
 
     setProgrammingListingServer(server);
   }, [initialMediaSourceId, mediaSources]);
-
-  useEffect(() => {
-    if (selectedServer?.type === Plex && plexLibraryChildren) {
-      if (
-        plexLibraryChildren.size > 0 &&
-        (!selectedLibrary || selectedLibrary.type !== Plex)
-      ) {
-        const initialLibrary = find(
-          plexLibraryChildren?.Directory,
-          ({ uuid }) => uuid === initialLibraryId,
-        );
-        setProgrammingListLibrary({
-          type: Plex,
-          view: {
-            type: 'library',
-            library: initialLibrary ?? plexLibraryChildren.Directory[0],
-          },
-        });
-        navigate({
-          search: {
-            libraryId: plexLibraryChildren.Directory[0].uuid,
-            mediaSourceId: selectedServer.id,
-          },
-        }).catch(console.error);
-      }
-      addKnownMediaForPlexServer(selectedServer.id, [
-        ...plexLibraryChildren.Directory,
-      ]);
-    } else if (selectedServer?.type === Jellyfin && jellyfinLibraries) {
-      if (
-        jellyfinLibraries.Items.length > 0 &&
-        (!selectedLibrary || selectedLibrary.type !== Jellyfin)
-      ) {
-        const view =
-          find(
-            jellyfinLibraries.Items,
-            ({ Id: id }) => id === initialLibraryId,
-          ) ?? sortBy(jellyfinLibraries.Items, sortJellyfinLibraries)[0];
-        setProgrammingListLibrary({
-          type: Jellyfin,
-          view,
-        });
-        navigate({
-          search: {
-            libraryId: view.Id,
-            mediaSourceId: selectedServer.id,
-          },
-        }).catch(console.error);
-      }
-      addKnownMediaForJellyfinServer(selectedServer.id, [
-        ...jellyfinLibraries.Items,
-      ]);
-    } else if (selectedServer?.type === Emby && embyLibraries) {
-      if (
-        embyLibraries.Items.length > 0 &&
-        (!selectedLibrary || selectedLibrary.type !== Emby)
-      ) {
-        setProgrammingListLibrary({
-          type: Emby,
-          view: first(sortBy(embyLibraries.Items, sortEmbyLibraries))!,
-        });
-      }
-      addKnownMediaForEmbyServer(selectedServer.id, embyLibraries.Items);
-    }
-  }, [
-    selectedServer,
-    plexLibraryChildren,
-    jellyfinLibraries,
-    selectedLibrary,
-    embyLibraries,
-    navigate,
-    initialLibraryId,
-  ]);
 
   /**
    * Load custom shows
@@ -225,72 +93,6 @@ export const ProgrammingSelector = ({
     [mediaSources, navigate],
   );
 
-  const onLibraryChange = useCallback(
-    (libraryUuid: string) => {
-      if (selectedServer?.type === Plex) {
-        if (libraryUuid === 'playlists' && plexPlaylists) {
-          setProgrammingListLibrary({
-            type: Plex,
-            view: { type: 'playlists', playlists: plexPlaylists },
-          });
-          return;
-        }
-
-        const library = knownMedia.getPlexMedia(selectedServer.id, libraryUuid);
-
-        if (library && isPlexDirectory(library)) {
-          setProgrammingListLibrary({
-            type: Plex,
-            view: { type: 'library', library },
-          });
-          navigate({
-            search: {
-              mediaSourceId: selectedServer.id,
-              libraryId: library.uuid,
-            },
-          }).catch(console.error);
-        }
-      } else if (selectedServer?.type === Jellyfin) {
-        const view = knownMedia.getMediaOfType(
-          selectedServer.id,
-          libraryUuid,
-          Jellyfin,
-        );
-        if (view) {
-          setProgrammingListLibrary({ type: Jellyfin, view });
-          navigate({
-            search: {
-              mediaSourceId: selectedServer.id,
-              libraryId: view.Id,
-            },
-          }).catch(console.error);
-        }
-      } else if (selectedServer?.type === Emby) {
-        const view = knownMedia.getMediaOfType(
-          selectedServer.id,
-          libraryUuid,
-          Emby,
-        );
-        if (view) {
-          setProgrammingListLibrary({ type: Emby, view });
-          navigate({
-            search: {
-              mediaSourceId: selectedServer.id,
-              libraryId: view.Id,
-            },
-          }).catch(console.error);
-        }
-      }
-    },
-    [
-      knownMedia,
-      navigate,
-      plexPlaylists,
-      selectedServer?.id,
-      selectedServer?.type,
-    ],
-  );
-
   const renderMediaSourcePrograms = () => {
     if (selectedLibrary) {
       switch (selectedLibrary.type) {
@@ -308,7 +110,6 @@ export const ProgrammingSelector = ({
               toggleOrSetSelectedProgramsDrawer={
                 toggleOrSetSelectedProgramsDrawer
               }
-              selectedGenre={selectedGenre}
             />
           );
         case Emby:
@@ -362,125 +163,15 @@ export const ProgrammingSelector = ({
 
     switch (selectedServer.type) {
       case Plex: {
-        const hasLibraries =
-          !isNil(plexLibraryChildren) &&
-          plexLibraryChildren.size > 0 &&
-          selectedPlexLibrary;
-
-        const libraryMenuItems = map(plexLibraryChildren?.Directory, (dir) => (
-          <MenuItem key={dir.key} value={dir.uuid}>
-            {dir.title}
-          </MenuItem>
-        ));
-
-        const playlistMenuItem = (
-          <MenuItem
-            key="playlists"
-            value="playlists"
-            disabled={plexPlaylistsLoading}
-          >
-            Playlists
-          </MenuItem>
-        );
-
-        return (
-          hasLibraries && (
-            <FormControl size="small" sx={{ minWidth: { sm: 200 } }}>
-              <InputLabel>Library</InputLabel>
-              <Select
-                label="Library"
-                value={
-                  selectedPlexLibrary.type === 'library'
-                    ? selectedPlexLibrary.library.uuid
-                    : 'playlists'
-                }
-                onChange={(e) => onLibraryChange(e.target.value)}
-              >
-                {libraryMenuItems}
-                {playlistMenuItem}
-              </Select>
-            </FormControl>
-          )
-        );
+        return <PlexLibrarySelector initialLibraryId={initialLibraryId} />;
       }
       case Jellyfin: {
-        return (
-          !isNil(jellyfinLibraries) &&
-          jellyfinLibraries.Items.length > 0 &&
-          selectedJellyfinLibrary && (
-            <FormControl size="small" sx={{ minWidth: { sm: 200 } }}>
-              <InputLabel>Library</InputLabel>
-              <Select
-                label="Library"
-                value={selectedJellyfinLibrary.Id}
-                onChange={(e) => onLibraryChange(e.target.value)}
-              >
-                {sortBy(jellyfinLibraries.Items, sortJellyfinLibraries).map(
-                  (lib) => (
-                    <MenuItem key={lib.Id} value={lib.Id}>
-                      {lib.Name}
-                    </MenuItem>
-                  ),
-                )}
-              </Select>
-            </FormControl>
-          )
-        );
+        return <JellyfinLibrarySelector initialLibraryId={initialLibraryId} />;
       }
       case Emby: {
-        return (
-          !isNil(embyLibraries) &&
-          embyLibraries.Items.length > 0 &&
-          selectedEmbyLibrary && (
-            <FormControl size="small" sx={{ minWidth: { sm: 200 } }}>
-              <InputLabel>Library</InputLabel>
-              <Select
-                label="Library"
-                value={selectedEmbyLibrary.Id}
-                onChange={(e) => onLibraryChange(e.target.value)}
-              >
-                {sortBy(embyLibraries.Items, sortEmbyLibraries).map((lib) => (
-                  <MenuItem key={lib.Id} value={lib.Id}>
-                    {lib.Name}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          )
-        );
+        return <EmbyLibrarySelector initialLibraryId={initialLibraryId} />;
       }
     }
-  };
-
-  const extractGenreName = (genre: JellyfinItem): string => genre.Name ?? '';
-
-  const renderGenreChoices = () => {
-    const genreList = jellyfinGenres?.Items ?? [];
-    return (
-      <FormControl size="small" sx={{ minWidth: { sm: 200 } }}>
-        <InputLabel>Genre</InputLabel>
-        <Select
-          label="Genre"
-          value={selectedGenre ?? ''}
-          onChange={(e) => {
-            const value = e.target.value;
-            setProgrammingGenre(value === '' ? undefined : value);
-          }}
-        >
-          <MenuItem value="">
-            <em>All Genres</em>
-          </MenuItem>
-          {genreList.map((genre) => {
-            const genreValue = extractGenreName(genre);
-            return (
-              <MenuItem key={genre.Id} value={genreValue}>
-                {capitalize(genreValue)}
-              </MenuItem>
-            );
-          })}
-        </Select>
-      </FormControl>
-    );
   };
 
   const hasAnySources = !isEmpty(mediaSources) || !isEmpty(customShows);
@@ -524,7 +215,6 @@ export const ProgrammingSelector = ({
           )}
 
           {renderLibraryChoices()}
-          {renderGenreChoices()}
           <ProgramViewToggleButton sx={{ ml: 'auto' }} />
         </Stack>
       </Box>
