+++ web/src/hooks/programming_controls/useAddProgramming.ts
@@ -1,7 +1,7 @@
 import { enumerateJellyfinItem } from '@/hooks/jellyfin/jellyfinHookUtil.ts';
 import { useKnownMedia } from '@/store/programmingSelector/selectors.ts';
 import { flattenDeep, map } from 'lodash-es';
-import { type MouseEventHandler, useState } from 'react';
+import { type MouseEventHandler, useCallback, useState } from 'react';
 import { match } from 'ts-pattern';
 import { Emby, Jellyfin, Plex } from '../../helpers/constants.ts';
 import { enumerateEmbyItem } from '../../helpers/embyUtil.ts';
@@ -21,95 +21,104 @@ export const useAddSelectedItems = () => {
   const selectedMedia = useStore((s) => s.selectedMedia);
   const [isLoading, setIsLoading] = useState(false);
 
-  const addSelectedItems: MouseEventHandler = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsLoading(true);
-    sequentialPromises(selectedMedia, (selectedMedia) =>
-      match(selectedMedia)
-        .returnType<Promise<AddedMedia[]>>()
-        .with({ type: Plex }, async (selected) => {
-          const media = knownMedia.getMediaOfType(
-            selected.serverId,
-            selected.id,
-            Plex,
-          );
+  const addSelectedItems: MouseEventHandler = useCallback(
+    (e) => {
+      e.preventDefault();
+      e.stopPropagation();
+      setIsLoading(true);
+      sequentialPromises(selectedMedia, (selectedMedia) =>
+        match(selectedMedia)
+          .returnType<Promise<AddedMedia[]>>()
+          .with({ type: Plex }, async (selected) => {
+            const media = knownMedia.getMediaOfType(
+              selected.serverId,
+              selected.id,
+              Plex,
+            );
 
-          if (!media) {
-            return [];
-          }
+            if (!media) {
+              return [];
+            }
 
-          return map(
-            await enumeratePlexItem(
+            return map(
+              await enumeratePlexItem(
+                apiClient,
+                selected.serverId,
+                selected.serverName,
+                media,
+              )(),
+              (item) => ({ media: item, type: Plex }),
+            );
+          })
+          .with({ type: Jellyfin }, async (selected) => {
+            const media = knownMedia.getMediaOfType(
+              selected.serverId,
+              selected.id,
+              Jellyfin,
+            );
+
+            if (!media) {
+              return [];
+            }
+
+            const items = await enumerateJellyfinItem(
               apiClient,
               selected.serverId,
               selected.serverName,
               media,
-            )(),
-            (item) => ({ media: item, type: Plex }),
-          );
+            )();
+
+            return map(items, (item) => ({ media: item, type: Jellyfin }));
+          })
+          .with({ type: Emby }, async (selected) => {
+            const media = knownMedia.getMediaOfType(
+              selected.serverId,
+              selected.id,
+              Emby,
+            );
+
+            if (!media) {
+              return [];
+            }
+
+            const items = await enumerateEmbyItem(
+              apiClient,
+              selected.serverId,
+              selected.serverName,
+              media,
+            )();
+
+            return map(items, (item) => ({ media: item, type: Emby }));
+          })
+          .with({ type: 'custom-show' }, (selected) => {
+            return Promise.resolve(
+              map(selected.programs, (p) => ({
+                type: 'custom-show',
+                customShowId: selected.customShowId,
+                totalDuration: selected.totalDuration,
+                program: p,
+              })),
+            );
+          })
+          .exhaustive(),
+      )
+        .then(flattenDeep)
+        .then(onAddSelectedMedia)
+        .then(() => {
+          clearSelectedMedia();
+          setIsLoading(false);
+          onAddMediaSuccess();
         })
-        .with({ type: Jellyfin }, async (selected) => {
-          const media = knownMedia.getMediaOfType(
-            selected.serverId,
-            selected.id,
-            Jellyfin,
-          );
-
-          if (!media) {
-            return [];
-          }
-
-          const items = await enumerateJellyfinItem(
-            apiClient,
-            selected.serverId,
-            selected.serverName,
-            media,
-          )();
-
-          return map(items, (item) => ({ media: item, type: Jellyfin }));
-        })
-        .with({ type: Emby }, async (selected) => {
-          const media = knownMedia.getMediaOfType(
-            selected.serverId,
-            selected.id,
-            Emby,
-          );
-
-          if (!media) {
-            return [];
-          }
-
-          const items = await enumerateEmbyItem(
-            apiClient,
-            selected.serverId,
-            selected.serverName,
-            media,
-          )();
-
-          return map(items, (item) => ({ media: item, type: Emby }));
-        })
-        .with({ type: 'custom-show' }, (selected) => {
-          return Promise.resolve(
-            map(selected.programs, (p) => ({
-              type: 'custom-show',
-              customShowId: selected.customShowId,
-              totalDuration: selected.totalDuration,
-              program: p,
-            })),
-          );
-        })
-        .exhaustive(),
-    )
-      .then(flattenDeep)
-      .then(onAddSelectedMedia)
-      .then(() => {
-        clearSelectedMedia();
-        setIsLoading(false);
-        onAddMediaSuccess();
-      })
-      .catch(console.error);
-  };
+        .catch(console.error);
+    },
+    [
+      apiClient,
+      knownMedia,
+      onAddMediaSuccess,
+      onAddSelectedMedia,
+      selectedMedia,
+    ],
+  );
 
   return { addSelectedItems, isLoading };
 };
